refactor(sync): move WebDAV client + connect settings under providers/webdav
